Reverted Cell.scala to not use roundToNeighborState and also added SilentCell
I reverted Cell to use the previous approach for tracking cell state.
Also, I found that changing the initialization of neighbors to use for
comprehension resulted in the compiler inferencing the type of the
neighbors Array to be Array[Any] because we don't want to add a cell to
its own neighbors array. Searching for a work around but for now, we
have a Silent Cell type which simply does not do anything when sent a
message.